Move the S3ObjectStore into the new storage lib
