+++ sbt_common/messaging/src/main/scala/uk/ac/wellcome/messaging/message/MessageWriter.scala
@@ -11,7 +11,7 @@ import uk.ac.wellcome.models.aws.S3Config
 
 import io.circe.Encoder
 import uk.ac.wellcome.messaging.sns.SNSWriter
-import uk.ac.wellcome.s3.{KeyPrefixGenerator, S3ObjectStore}
+import uk.ac.wellcome.storage.s3.{KeyPrefixGenerator, S3ObjectStore}
 import uk.ac.wellcome.utils.JsonUtil._
 
 import scala.concurrent.{blocking, Future}
